+++ src/parser/dissectors/link/ethernet/types.ts
@@ -1,0 +1,9 @@
+import type { NetworkLayer } from "../../network/types";
+
+export type EthernetFrame = {
+  type: "ethernet";
+  destMac: string;
+  srcMac: string;
+  etherType: number;
+  payload: NetworkLayer
+}