+++ src/parser/dissectors/link/types.ts
@@ -1,0 +1,10 @@
+import type { EthernetFrame } from "./ethernet/types"
+
+export type LinkLayer = EthernetFrame | UnknownLinkLayer;
+
+
+export type UnknownLinkLayer = {
+  type: "unknown-link",
+  raw: Uint8Array
+}
+