+++ src/parser/dissectors/transport/udp/types.ts
@@ -1,0 +1,7 @@
+export type UDPPacket = {
+  type: "udp",
+  srcPort: number;
+  dstPort: number;
+  length: number;
+  payload: Uint8Array
+}