+++ src/parser/utils/readIP.ts
@@ -1,0 +1,13 @@
+import type { BinaryReader } from "../core/BinaryReader"
+
+
+function readIP(reader: BinaryReader) {
+  return [
+    reader.readUInt8(),
+    reader.readUInt8(),
+    reader.readUInt8(),
+    reader.readUInt8()
+  ].join(".")
+}
+
+export {readIP}