+++ src/parser/utils/toMAC.ts
@@ -1,0 +1,3 @@
+export function toMAC(bytes: Uint8Array) {
+  return Array.from(bytes).map(b => b.toString(16).padStart(2, "0")).join(":")
+}