+++ CHANGELOG.md
@@ -5,3 +5,8 @@ the stuff changing duh dawg
 - initialized project
 - created home page and drag and drop upload
 - added workspace header
+
+### 28-03-2026
+- created a group of dissectors to parse ethernet link layer packets (ethernet -> IP -> TCP/UDP)
+- types of all the packets and frames
+- file structure to manage parsers of all 4 layers